Update SaveReminderFragment.kt
this function handles the user's response to the permissions.permissions can be denied if grantResult array is empty or the user actually denied the permissions or the device is running on api Q or higher and the user denied the background permission.
if permissions is denied show the user a snackbar explaining the user that the application would be useless without these permissions.
if permissions is granted then call checkDeviceLocationSettingsAndStartGeofence() method